Set alarm correctly if interval is changed in Preferences. Change int to long, because longest interval does not fit into int.
git-svn-id: https://openintents.googlecode.com/svn/trunk@1373 72b678ce-9140-0410-bee8-679b907dd61a

+++ UpdateChecker/src/org/openintents/updatechecker/UpdateCheckService.java
@@ -156,8 +156,8 @@ public class UpdateCheckService extends Service {
 	}
 
 	private void setAlarm(Intent intent, int startId) {
-		int interval = intent.getIntExtra(EXTRA_INTERVAL, -1);
-		int firstinterval = intent.getIntExtra(EXTRA_FIRST_INTERVAL, -1);
+		long interval = intent.getLongExtra(EXTRA_INTERVAL, -1);
+		long firstinterval = intent.getLongExtra(EXTRA_FIRST_INTERVAL, -1);
 		if (firstinterval < 0) {
 			firstinterval = interval;
 		}
